requirements.txt
streamlit
pandas
scikit-learn
plotly
pillow
numpy
protobuf
